[ADDED] Particle module [CHANGED] moved vulkan_helper.hpp to public so particles can access it [FIXED] wrong includes after moving particles and vulkan_helper
