+++ engine/include/old_engine.hpp
@@ -2,7 +2,7 @@
 
 #include "mesh.hpp"
 #include "module_interface.hpp"
-#include "particles/particle_interface.hpp"
+#include "particle_interface.hpp"
 #include "performance_tracker.hpp"
 
 #include <memory>
