+++ engine/particles/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_library(Particles)
+module_default_init(Particles)
+
+target_link_libraries(Particles
+        PUBLIC Core
+        PUBLIC Renderer
+        PUBLIC OldEngine
+)
