+++ engine/particles/private/emitter_component.cpp
@@ -1,0 +1,1 @@
+#include "emitter_component.hpp"
