+++ engine/particles/public/particle_util.hpp
