+++ engine/renderer/CMakeLists.txt
@@ -5,4 +5,5 @@ target_link_libraries(Renderer
         PUBLIC Core
         PUBLIC Application
         PUBLIC OldEngine
+        PUBLIC Particles
 )
