--- engine/renderer/private/particles/emitter_component.cpp
@@ -1,1 +1,0 @@
-#include "particles/emitter_component.hpp"
