--- engine/renderer/private/particles/particle_util.cpp
@@ -1,1 +1,0 @@
-#include "particles/particle_util.hpp"
