+++ engine/renderer/private/pipelines/lighting_pipeline.cpp
@@ -1,6 +1,6 @@
 #include "pipelines/lighting_pipeline.hpp"
 
-#include "../vulkan_helper.hpp"
+#include "vulkan_helper.hpp"
 #include "bloom_settings.hpp"
 #include "camera.hpp"
 #include "gpu_scene.hpp"
