+++ engine/source/engine.cpp
@@ -19,9 +19,9 @@
 #include "input_manager.hpp"
 #include "model_loader.hpp"
 #include "old_engine.hpp"
-#include "particles/emitter_component.hpp"
-#include "particles/particle_interface.hpp"
-#include "particles/particle_util.hpp"
+#include "emitter_component.hpp"
+#include "particle_interface.hpp"
+#include "particle_util.hpp"
 #include "physics_module.hpp"
 #include "pipelines/debug_pipeline.hpp"
 #include "profile_macros.hpp"
